+++ apps/web/src/lib/access-control/page-permission-helper.ts
@@ -1,5 +1,6 @@
 
-import type { RoleType } from '@/schema/inventory/cloud-service-report/model';
+
+import type { RoleType } from '@/schema/identity/role/type';
 
 import type {
     PagePermissionMap,
@@ -17,6 +18,7 @@ import type { MenuId } from '@/lib/menu/config';
 import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
 
 import type { LNBItem, LNBMenu } from '@/common/modules/navigations/lnb/type';
+
