+++ apps/web/src/schema/identity/role/type.ts
@@ -1,0 +1,3 @@
+import type { ROLE_TYPE } from '@/schema/identity/role/constant';
+
+export type RoleType = typeof ROLE_TYPE[keyof typeof ROLE_TYPE];
