+++ apps/web/src/services/administration/pages/RoleUpdatePage.vue
@@ -35,7 +35,7 @@ import { PHeading, PButton } from '@spaceone/design-system';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 import { SpaceRouter } from '@/router';
-import type { RoleModel } from '@/schema/inventory/cloud-service-report/model';
+import type { RoleModel } from '@/schema/identity/role/model';
 import { i18n } from '@/translations';
 
 import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
