+++ apps/web/src/services/administration/store/policy-page-store.ts
@@ -5,7 +5,7 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 
 import type { ListResponse } from '@/schema/_common/model';
-import type { PolicyModel } from '@/schema/identity/user-group/model';
+import type { PolicyModel } from '@/schema/identity/policy/model';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
