+++ apps/web/src/store/modules/user/getters.ts
@@ -1,6 +1,6 @@
 import type { Getter } from 'vuex';
 
-import type { RoleType } from '@/schema/inventory/cloud-service-report/model';
+import type { RoleType } from '@/schema/identity/role/type';
 
 import { languages } from '@/store/modules/user/config';
 
@@ -13,6 +13,7 @@ import {
 
 
 import type { UserState } from './type';
+
 
 export const isDomainOwner = (state: UserState): boolean => state.userType === 'DOMAIN_OWNER';
 export const languageLabel = (state: UserState): string => languages[state.language as string] || state.language;
